test(community): move delegated-IPNS benchmark to scripts/, stabilize failure tests
Address CodeRabbit feedback on the delegated-IPNS PR:

- Move the env-gated BENCH_IPNS timing benchmark out of the vitest suite into
  scripts/bench-delegated-ipns.js (run with the test server up). Running the
  script is the opt-in, so the brittle Boolean(BENCH_IPNS) truthy gate (which
  enabled on "0"/"false") is gone. Update docs/protocol/delegated-ipns.md to
  point at the script.
- Route the five expected-failure load tests through createCommunity()+update()
  via a new loadCommunityExpectingError helper instead of the one-shot
  getCommunity(), per the coding guideline, removing CI flake risk from
  one-shot transport timing.
